docs(README): update links to preferred syntax files
This updates the links to two plugins:

1. java-syntax.vim
2. rust.vim

java-syntax.vim  hasn't had a commit in a few months, and is pending a
PR from me which will enable syntax highlighting of delimiters. For the
time being, it would be better to use that.

rust.vim has made the decision to not allow highlighting of user types,
and while I understand their reasoning, I completely disagree. For that
reason I will maintain the fork and merge any changes from upstream.