ref(colors): add NONE constant
The appearance of a lone `{}` was perhaps confusing to some. Behind the
scenes, whenever an attribute is missing from a table, it will be
defaulted to 'None'. Transiently, in using `{}` we are able to shortcut
directly to these defaults.

However, it does not read well. This commit introduces a constant that
makes it clear what is happening.