feat(colors): add more TOML default links
For some reason, `tomlFloat` was linked to `Number`. There is a `Float`
highlight group dedicated for that.

Additionally, the `tomlDate` group was linked to `Constant`. I don't
think that's necessarily a fitting group, since a date is a literal,
and while literals _are_ constants it does not seem fitting. I gave it
the `Special` link instead, since `Special` is supposed to be used
whenever there is an "other" type in a language.