docs: document overriding groups from init.vim
This adds instructions on how to override highlight groups from `init.vim`,
as it seems this cannot be done with your average Neovim colorscheme.
Rather than just using `:hi!` in the `init.vim` (as one might be tempted
to do), it describes a process of adding the commands to an autocommand
event which will run them when the time is right.

SEE #9