fix(highlite): highlight groups being cleared
This was a step I took from the plugin this colorscheme is forked from
(romainl/vim-rnb) but it turns out that it causes problems and doesn't
actually solve anything.

Taking out this line fixed a highlight group being cleared when it
should not have been, although interestingly not _all_ highlight groups
are being cleared consistently. So its just introducing inconsistency
into this plugin which I do not want.

SEE glepnir/galaxyline.nvim#88