fix(export): report error to user if automatic directory detection does not work
If a user does not specify the directory while exporting a colorscheme,
nvim-highlite will attempt to resolve the proper directory. However, it
seems that I did not implement proper error handling while writing this
section of the code.

Now, if this happens, users will be informed that automatic resolution
has failed, and that the program is falling back to the CWD.