fix(groups): `Comment` group derived from `text_contrast_bg_low` color
It seems that, instead of using the dedicated `comment` color in the
palette, it was derived from the `text_contrast_bg_low` color directly.

This commit amends that mistake: it should not be seen as a breaking
change, as it was always meant to be used this way, and by default,
`comment` _is_ `text_contrast_bg_low`.